Fix clang compilation warning:
khtml/src/xml/dom_stringimpl.h:60:13: warning: cast from 'char *'
to 'QChar *' increases required alignment from 1 to 2 [-Wcast-align
        s = (QChar*) new cha sizeof(QChar)*( havestr ? len : 1 ) ;
            ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Note that this line of code is pretty odd I have to say. It this
supposed to be an optimization to prevent the default ctor/dtor of
QChar to be called? Is it really worth it here?

REVIEW: 116544