Fix potential mem leak in CSS parsing.
Fix potential memleak (and possibly CSS mis-parsing) caught by Coverity,
CID 257928.

This one is hard to follow, but in essence if we parse through the list
of tokens being examined by parseBackgroundProperty, and end up with a
*list* in values and a single value2 at the end, what actually gets
returned is the list of values and the (empty) list values2. value2 is
non-zero and leaked in this scenario.

I try to fix by matching the logic used at the end of the procedure with
the code that migrates a value from value -> values: if we switch from a
single entry to holding a list, we do so for value2/values2 as well.