Correctly use QCache::insert.
QCache documents that insert can delete the pointer you pass to it
literally as soon as the call completes, so we can't rely on the pointer
being valid after ::insert() (although that's pretty much never an issue
in practice).

Coverity duly warns about this (CID 259759), so fix by hoisting a copy
of the pixmap out first.

QCache::object also warns that the return value can be deleted at any
time, but this code is the only user of that cache and is not otherwise
thread-safe, so that seems like a non-issue.