Error page, correctly load the image (with a real URL)
Summary:
The local address to the image is injected into an HTML page (the
error page). This address needs to be a proper URL with file:///,
not just the absolute path to the image file - otherwise
KIO::SimpleJobPrivate::simpleJobInit complains with "Invalid URL".

Test Plan:
Compiles, and when trying to load an invalid address in Konqueror
with the KHTML part, the warning picture on the top-left corner is
now visible.

Reviewers: mpyne

Reviewed By: mpyne

Subscribers: mpyne, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D5104