Generate kentities gperf output at build time
Use gperf to generate the hash-based lookup for the entities; this
replaces the static generated file in the sources.

This also gets rid of the hack to cope with the duplicate symbols,
as the gperf output was included in two places: instead, add a simple
wrapper function so the gperf output is built only once, and the
internals of the kentities struct are not exposed anymore.

As result, kentities.gperf now is used as-is from kcodecs, so it is
way easier to get improvements from that version.